Update draft-ietf-tcpm-rfc8312bis.md
This PR is intended to amend the text to align with my understanding of the RFC-series. This normatively  would require ssthresh to be set from flight_size; however, this is explicitly relaxed in RFC7661, so I suggest wording should be careful around this topic. I retained the original text observation that in fact some cubic implementations are known to set ssthresh from cwnd.